move 'benchpark setup' to its own cmd module; refactor some resources into separate modules that are used by setup and main
